+++ src/main/java/me/nikl/lmgtfy/Language.java
@@ -9,23 +9,22 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.InputStreamReader;
 import java.io.UnsupportedEncodingException;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Created by nikl on 24.10.17.
  *
- * Cleaned up super Language class from GameBox...
  * Hold messages and handle the language files.
+ * Default files are handled in {@link me.nikl.lmgtfy.util.FileUtil}.
  */
 public class Language {
 
-    protected Main plugin;
+    private Main plugin;
 
-    protected File languageFile;
+    private File languageFile;
 
-    protected FileConfiguration defaultLanguage;
-    protected FileConfiguration language;
+    private FileConfiguration defaultLanguage;
+    private FileConfiguration language;
 
     public String PREFIX = "["+ChatColor.DARK_AQUA+"LMGTFY"+ChatColor.RESET+"]";
     public String NAME = ChatColor.DARK_AQUA+"LMGTFY"+ChatColor.RESET;
@@ -51,7 +50,7 @@ public class Language {
     public String CMD_MISSING_QUERY, CMD_NO_PERM;
 
 
-    public Language(Main plugin){
+    Language(Main plugin){
         this.plugin = plugin;
 
         getLangFile(plugin.getConfig());
@@ -120,7 +119,7 @@ public class Language {
      * 'lang_xx.yml': will try to load the given file inside the namespaces language folder
      * @param config
      */
-    protected void getLangFile(FileConfiguration config) {
+    private void getLangFile(FileConfiguration config) {
         // load default language
         try {
             String defaultLangName = "language/lang_en.yml";
@@ -171,60 +170,6 @@ public class Language {
 
         return;
     }
-
-
-    /**
-     * Find all string messages that are missing in the language file.
-     *
-     * This method compares all message keys that hold a String in the default english
-     * file with all set keys in the used language file. All missing keys are
-     * collected and returned.
-     *
-     * @return list of all missing keys (can be empty list)
-     */
-    public List<String> findMissingStringMessages(){
-
-        List<String> toReturn = new ArrayList<>();
-
-        if(defaultLanguage.equals(language)) return toReturn;
-
-        for(String key : defaultLanguage.getKeys(true)){
-            if(defaultLanguage.isString(key)){
-                if(!language.isString(key)){
-                    // there is a message missing
-                    toReturn.add(key);
-                }
-            }
-        }
-        return toReturn;
-    }
-
-    /**
-     * Find all string messages that are missing in the language file.
-     *
-     * This method compares all message keys that hold a list in the default english
-     * file with all set keys in the used language file. All missing keys are
-     * collected and returned.
-     *
-     * @return list of all missing keys (can be empty list)
-     */
-    public List<String> findMissingListMessages(){
-
-        List<String> toReturn = new ArrayList<>();
-
-        if(defaultLanguage.equals(language)) return toReturn;
-
-        for(String key : defaultLanguage.getKeys(true)){
-            if (defaultLanguage.isList(key)){
-                if(!language.isList(key)){
-                    // there is a list missing
-                    toReturn.add(key);
-                }
-            }
-        }
-        return toReturn;
-    }
-
 
     /**
      * Load list messages from the language file